Subo los ultimos cambios. Considerar que el archivo webConfig está con persmisos de acceso a todos los usuarios para poder probarlo. Siguen habiendo errores

+++ SI-ASA-ENTIDADESv1/Docente.cs
@@ -10,6 +10,6 @@ namespace SI_ASA_ENTIDADESv1
         public Persona docente { get; set; }
         public int legajo { get; set; }
         public Horario horarioTrabajo { get; set; }
-        public int salario { get; set; }
+        public float salario { get; set; }
     }
 }
